requirements.txt
# 工业缺陷检测系统 - Python 依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
#pytorch-cuda  # 可选，仅 GPU 用户需要

# YOLO11 (Ultralytics)
ultralytics>=8.3.0

# 数据处理
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0

# 数据增强
albumentations>=1.1.0

# 配置管理
pyyaml>=5.4.0

# 推理引擎
onnx>=1.12.0
onnxruntime>=1.14.0  # CPU
# onnxruntime-gpu>=1.14.0  # GPU 推理

# 进度条
tqdm>=4.60.0

# 其他工具
scikit-learn>=0.24.0
matplotlib>=3.3.0
tensorboard>=2.5.0

# 可选：模型优化
onnx-simplifier>=0.4.0

# GUI界面
PySide6>=6.5.0
